Remove glyphicons and old MT client.  Merge wiki docs.  Add apache headers.
Changed the icons to remove need for Glyphicons.  Now all icons are
provided by Font Awesone, which is installed as an npm dependency.

Retired old MTConnect client directory.  It was no longer used because
the necessary code is now in the main client directory.

The twelve pages in the package GitHub wiki are now in a new docs/api
directory, and linked into the root index file.  Keep all docs in the
package.  Moved outdated API.md to the subdir as older.md

All source files now have the recommended apache license header.  Also
minor edits to package.json to properly identify the package.

Signed-off-by: David Loffredo <loffredo@steptools.com>